api/requirements.txt
fastapi
pydantic
passlib
pymongo
python-jose
uvicorn
google.generativeai
bcrypt
httpx
python-multipart
pandas

